Navigation周りをリファクタリング
