+++ app/src/main/java/com/example/cleanarchitecturesample/feature_app/presentation/navigation/NavDestination.kt
@@ -5,7 +5,7 @@ sealed class NavDestination(val route: String) {
     //region route
     data object Splash : NavDestination("splash")
     data object AuthGraph : NavDestination("auth_graph")
-    data object AppRoute : NavDestination("app_route")
+    data object AppContentsGraph : NavDestination("app_contents_graph")
     //endregion
 
     //region AuthGraph
@@ -19,6 +19,12 @@ sealed class NavDestination(val route: String) {
     data object StatisticsGraph : NavDestination("statistics_graph")
     data object StatisticsUser : NavDestination("statistics_user")
     data object StatisticsAll : NavDestination("statistics_all")
-    data class Product(val key: String = "productId") : NavDestination("product/{$key}")
+    data object Product : NavDestination("product/{${PRODUCT_ID}}") {
+        fun passId(id: String): String = route.replace("{$PRODUCT_ID}", id)
+    }
     //endregion
+
+    companion object {
+        const val PRODUCT_ID = "productId"
+    }
 }
